+++ src/providers/asdf.js
@@ -39,6 +39,14 @@ module.exports = {
     }
   },
 
+  installCommand(version) {
+    return `asdf install nodejs ${version}`;
+  },
+
+  uninstallCommand(version) {
+    return `asdf uninstall nodejs ${version}`;
+  },
+
   binDir() {
     return null;
   }