+++ src/providers/fnm.js
@@ -54,6 +54,14 @@ module.exports = {
     return fnm(`default ${version}`);
   },
 
+  installCommand(version) {
+    return `fnm install ${version}`;
+  },
+
+  uninstallCommand(version) {
+    return `fnm uninstall ${version}`;
+  },
+
   binDir(version) {
     const base = path.join(versionsDir(), `v${version}`, 'installation');
     return U.IS_WIN ? base : path.join(base, 'bin');